+++ src/main/java/com/codecool/fplbackend/repository/UserRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User,Long> {
 
-    Optional<User> findUserByGithubId(Long id);
+    Optional<User> findUserByGithubId(Integer id);
 
 }
